Tiny suggestion for editable installations
In the section Installation, it gives an example of doing a local
editable installation, and in the section Development and Testing, it
gives an example of installing the package for development situations.
Although it's true that the instructions as given will work after a
fashion, the approach of using `pip install -e` is both more common
for editable installs, and also arguably leads to more intuitive
behavior when developing Python packages.